+++ bin/wsc.rs
@@ -29,6 +29,10 @@ fn main() -> Result<(), pa::Error> {
     let main = parsed.table.get("main").unwrap();
     let init = parsed.init;
 
+    for (name, op) in parsed.table.clone()  {
+        println!("{:?}", name);
+    }
+
     if args.is_present("print") {
         let composition = render(main, init);
         to_wav(composition, filename.unwrap().to_string());
