+++ songs/std.socool
@@ -16,6 +16,6 @@ fade_out = {
 }
 
 main = {
-	AsIs, 	
+	AsIs 	
 }
